twolip: Provide android.frameworks.sensorservice@1.0 on vendor
Following the commit ("tuli: Provide radio, keymaster and secure_element HAL on vendor"),
starting with Android 12 most HAL libraries are not included in VNDK anymore.
They need to be explicitly pulled in to `/vendor` for odm blobs to be allowed to access them:

    QCamera : <MCI><ERROR> mm_camera_load_shim_lib: 3306: dlopen failed
    with error dlopen failed: library
    "android.frameworks.sensorservice@1.0.so" not found: needed by
    /odm/lib/libmmcamera2_stats_modules.so in namespace (default)

android.frameworks.sensorservice@1.0 is used by at least mm-camera framework.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>